Broke out checkGoalObstruction and refactored the parameters to pass in the board. #28
